Add notebook configuration This enables configuration of notebooks with a notebook.yaml file, which generated with `jw new`. Each notebook directory now requires a notebook.yaml file to do any operations on it.
The post_path and metadata options enable a lot of configuration
of notebook structure. Now it can be flexible to a lot more than
my style of micro-journaling.

This commit also includes a tiny YAML parser for the very little that
needs to be parsed in notebook.yaml. It only supports basic `key: value`
strings.

TODO still:
- More code that may need to be ported (that currently use notebook.id)
- More flexibility in the configuration (specifically with metadata),
  requiring a better YAML parser.
- Not requiring a temporary file when the file name is not tied to the
  post's text